% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_final_data.R
\docType{data}
\name{df_final}
\alias{df_final}
\title{an example of the result of doIntegration}
\format{An object of data.frame.}
\usage{
data(df_final)
}
\description{
a dataframe contatins log2FoldChange of RNA-Seq and ChIP-Seq and the Z score
 for each mark
}
\examples{
data(df_final)
}
\keyword{datasets}
