% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoter_data.R
\docType{data}
\name{promoter}
\alias{promoter}
\title{an example of promoter region}
\format{An object of GRanges.}
\usage{
data(promoter)
}
\description{
an example of promoter region
}
\examples{
data(promoter)
}
\keyword{datasets}
