% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_data.R
\docType{data}
\name{res}
\alias{res}
\title{an example of the result of doMatch}
\format{An object of list.}
\usage{
data(res)
}
\description{
a list with the following three items.
 1. res.rna, a data frame contains RNA-Seq counts
 2. res.chip, a data frame contains ChIP-Seq counts
 3. matched.data, a dataframe contains matched RNA-Seq counts and ChIP-Seq
 counts
}
\examples{
data(res)
}
\keyword{datasets}
