% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chip_meta_data.R
\docType{data}
\name{test_chip_meta}
\alias{test_chip_meta}
\title{meta data of preprocessed ChIP-Seq data}
\format{An object of data.frame.}
\usage{
data(test_chip_meta)
}
\description{
The ChIP-Seq meta data.frame at least three columns:
1. mark: the mark of histone modifications (e.g. H3K4me3 or H3K27ac).
2. condition: identifier of the condition to which each sample belongs.
3. files: the exact address of the aligned bam files.
}
\examples{
data(test_chip_meta)
}
\keyword{datasets}
