% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_rna_meta_data.R
\docType{data}
\name{test_rna_meta}
\alias{test_rna_meta}
\title{meta data of preprocessed RNA-Seq data}
\format{An object of data.frame.}
\usage{
data(test_rna_meta)
}
\description{
The RNA-Seq meta data.frame at least two columns:
1. condition: identifier of the condition to which each sample belongs.
2. files: the exact address of the files contains the tsv file
which is the output of RNA-Seq preprocessed with Kallisto.
}
\examples{
data(test_rna_meta)
}
\keyword{datasets}
